Moves BundleSubRequest constant
